Refactoring WAI tests
