+++ src/systems/a320_systems/src/pneumatic/wing_anti_ice.rs
@@ -49,9 +49,9 @@ impl WingAntiIceValveSignal {
         Self { target_open_amount }
     }
 
-    pub fn new_open() -> Self {
-        Self::new(Ratio::new::<percent>(100.))
-    }
+    // pub fn new_open() -> Self {
+    //     Self::new(Ratio::new::<percent>(100.))
+    // }
 
     pub fn new_closed() -> Self {
         Self::new(Ratio::new::<percent>(0.))
@@ -125,9 +125,9 @@ impl WingAntiIceValveController {
         self.controller_signals_on
     }
 
-    pub fn controller_valve_pid_output(&self) -> f64 {
-        self.valve_pid.output()
-    }
+    // pub fn controller_valve_pid_output(&self) -> f64 {
+    //     self.valve_pid.output()
+    // }
 
     pub fn update(
         &mut self,
@@ -296,7 +296,6 @@ pub struct WingAntiIceComplex {
     wai_system_has_fault: bool,
     wai_system_on: bool,
     wai_selected: bool,
-    wai_shed: bool,
     wai_high_pressure: [bool; NUM_OF_WAI],
     wai_low_pressure: [bool; NUM_OF_WAI],
     wai_bleed_pressurised: [bool; NUM_OF_WAI],
@@ -319,8 +318,8 @@ pub struct WingAntiIceComplex {
 impl WingAntiIceComplex {
     const WAI_MIN_PRESSURE: f64 = 1.; //BAR
     const WAI_MAX_PRESSURE: f64 = 2.1; //BAR
-    const WAI_EXHAUST_SPEED: f64 = 0.1785;
-    const WAI_VALVE_TRANSFER_SPEED: f64 = 0.023;
+    const WAI_EXHAUST_SPEED: f64 = 0.17;
+    const WAI_VALVE_TRANSFER_SPEED: f64 = 1.3;
     // Each WAI duct is made of
     // Flow Trimming Restrictor 47mm diameter
     // Lagged Supply Duct 50mm diameter
@@ -372,7 +371,6 @@ impl WingAntiIceComplex {
             wai_system_has_fault: false,
             wai_system_on: false,
             wai_selected: false,
-            wai_shed: false,
             wai_high_pressure: [false, false],
             wai_low_pressure: [false, false],
             wai_bleed_pressurised: [false, false],
@@ -447,6 +445,18 @@ impl WingAntiIceComplex {
 
     pub fn wai_consumer_temperature(&self, number: usize) -> ThermodynamicTemperature {
         self.wai_consumer[number].temperature()
+    }
+
+    pub fn wai_mass_flow(&self, number: usize) -> MassRate {
+        self.wai_exhaust[number].fluid_flow()
+    }
+
+    pub fn wai_timer(&self, number: usize) -> Duration {
+        self.valve_controller[number].get_timer()
+    }
+
+    pub fn wai_valve_controller_on(&self, number: usize) -> bool {
+        self.valve_controller[number].controller_signals_on()
     }
 
     // This is where the action happens
@@ -525,7 +535,8 @@ impl WingAntiIceComplex {
         }
 
         self.wai_system_has_fault = has_fault;
-        self.wai_system_on = (num_of_on == NUM_OF_WAI); // && !has_fault;
+        // TODO Check what is the input to system_on
+        self.wai_system_on = num_of_on == NUM_OF_WAI; // && !has_fault;
         self.wai_selected = wai_mode == WingAntiIcePushButtonMode::On;
     }
 }
@@ -561,1248 +572,5 @@ impl SimulationElement for WingAntiIceComplex {
         writer.write(&self.wai_right_high_pressure_id, self.wai_high_pressure[1]);
         writer.write(&self.wai_left_low_pressure_id, self.wai_low_pressure[0]);
         writer.write(&self.wai_right_low_pressure_id, self.wai_low_pressure[1]);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use systems::{
-        air_conditioning::{
-            acs_controller::{Pack, PackFlowController},
-            AirConditioningSystem, PackFlowControllers, ZoneType,
-        },
-        electrical::{test::TestElectricitySource, ElectricalBus, Electricity},
-        engine::leap_engine::LeapEngine,
-        pneumatic::{EngineState, TargetPressureTemperatureSignal},
-        pressurization::PressurizationOverheadPanel,
-        shared::{
-            ApuBleedAirValveSignal, Cabin, ControllerSignal, ElectricalBusType, ElectricalBuses,
-            EmergencyElectricalState, EngineBleedPushbutton, EngineCorrectedN1,
-            EngineFirePushButtons, EngineStartState, GroundSpeed, HydraulicColor,
-            InternationalStandardAtmosphere, LgciuWeightOnWheels, MachNumber, PackFlowValveState,
-            PneumaticBleed, PneumaticValve, PotentialOrigin,
-        },
-        simulation::{
-            test::{SimulationTestBed, TestBed},
-            Aircraft, SimulationElement,
-        },
-    };
-
-    use std::time::Duration;
-
-    use uom::si::{
-        length::foot,
-        mass_rate::kilogram_per_second,
-        pressure::{pascal, psi},
-        thermodynamic_temperature::degree_celsius,
-        velocity::knot,
-    };
-
-    use crate::{
-        hydraulic::A320Hydraulic,
-        systems::simulation::test::{ReadByName, WriteByName},
-        A320Pneumatic, A320PneumaticOverheadPanel,
-    };
-
-    struct TestAirConditioning {
-        a320_air_conditioning_system: AirConditioningSystem<3>,
-
-        adirs: TestAdirs,
-        lgciu: TestLgciu,
-        pressurization: TestPressurization,
-        pressurization_overhead: PressurizationOverheadPanel,
-    }
-    impl TestAirConditioning {
-        fn new(context: &mut InitContext) -> Self {
-            let cabin_zones: [ZoneType; 3] =
-                [ZoneType::Cockpit, ZoneType::Cabin(1), ZoneType::Cabin(2)];
-
-            Self {
-                a320_air_conditioning_system: AirConditioningSystem::new(context, cabin_zones),
-
-                adirs: TestAdirs::new(),
-                lgciu: TestLgciu::new(true),
-                pressurization: TestPressurization::new(),
-                pressurization_overhead: PressurizationOverheadPanel::new(context),
-            }
-        }
-        fn update(
-            &mut self,
-            context: &UpdateContext,
-            engines: [&impl EngineCorrectedN1; 2],
-            engine_fire_push_buttons: &impl EngineFirePushButtons,
-            pneumatic: &(impl EngineStartState + PackFlowValveState + PneumaticBleed),
-            pneumatic_overhead: &impl EngineBleedPushbutton,
-        ) {
-            self.a320_air_conditioning_system.update(
-                context,
-                &self.adirs,
-                engines,
-                engine_fire_push_buttons,
-                pneumatic,
-                pneumatic_overhead,
-                &self.pressurization,
-                &self.pressurization_overhead,
-                [&self.lgciu; 2],
-            );
-        }
-    }
-    impl PackFlowControllers<3> for TestAirConditioning {
-        fn pack_flow_controller(&self, pack_id: Pack) -> PackFlowController<3> {
-            self.a320_air_conditioning_system
-                .pack_flow_controller(pack_id)
-        }
-    }
-    impl SimulationElement for TestAirConditioning {
-        fn accept<V: SimulationElementVisitor>(&mut self, visitor: &mut V) {
-            self.a320_air_conditioning_system.accept(visitor);
-
-            visitor.visit(self);
-        }
-    }
-
-    struct TestAdirs {
-        ground_speed: Velocity,
-    }
-    impl TestAdirs {
-        fn new() -> Self {
-            Self {
-                ground_speed: Velocity::new::<knot>(0.),
-            }
-        }
-    }
-    impl GroundSpeed for TestAdirs {
-        fn ground_speed(&self) -> Velocity {
-            self.ground_speed
-        }
-    }
-
-    struct TestPressurization {
-        cabin_pressure: Pressure,
-    }
-    impl TestPressurization {
-        fn new() -> Self {
-            Self {
-                cabin_pressure: Pressure::new::<pascal>(101315.),
-            }
-        }
-    }
-    impl Cabin for TestPressurization {
-        fn altitude(&self) -> Length {
-            Length::new::<foot>(0.)
-        }
-
-        fn pressure(&self) -> Pressure {
-            self.cabin_pressure
-        }
-    }
-
-    struct TestLgciu {
-        compressed: bool,
-    }
-    impl TestLgciu {
-        fn new(compressed: bool) -> Self {
-            Self { compressed }
-        }
-    }
-    impl LgciuWeightOnWheels for TestLgciu {
-        fn left_and_right_gear_compressed(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            self.compressed
-        }
-        fn right_gear_compressed(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            true
-        }
-        fn right_gear_extended(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            false
-        }
-        fn left_gear_compressed(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            true
-        }
-        fn left_gear_extended(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            false
-        }
-        fn left_and_right_gear_extended(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            false
-        }
-        fn nose_gear_compressed(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            true
-        }
-        fn nose_gear_extended(&self, _treat_ext_pwr_as_ground: bool) -> bool {
-            false
-        }
-    }
-
-    struct TestApu {
-        bleed_air_valve_signal: ApuBleedAirValveSignal,
-        bleed_air_pressure: Pressure,
-        bleed_air_temperature: ThermodynamicTemperature,
-    }
-    impl TestApu {
-        fn new() -> Self {
-            Self {
-                bleed_air_valve_signal: ApuBleedAirValveSignal::new_closed(),
-                bleed_air_pressure: Pressure::new::<psi>(14.7),
-                bleed_air_temperature: ThermodynamicTemperature::new::<degree_celsius>(15.),
-            }
-        }
-
-        fn update(&self, bleed_valve: &mut impl ControllablePneumaticValve) {
-            bleed_valve.update_open_amount::<ApuBleedAirValveSignal, Self>(self);
-        }
-
-        fn set_bleed_air_pressure(&mut self, pressure: Pressure) {
-            self.bleed_air_pressure = pressure;
-        }
-
-        fn set_bleed_air_temperature(&mut self, temperature: ThermodynamicTemperature) {
-            self.bleed_air_temperature = temperature;
-        }
-
-        fn set_bleed_air_valve_signal(&mut self, signal: ApuBleedAirValveSignal) {
-            self.bleed_air_valve_signal = signal;
-        }
-    }
-    impl ControllerSignal<ApuBleedAirValveSignal> for TestApu {
-        fn signal(&self) -> Option<ApuBleedAirValveSignal> {
-            Some(self.bleed_air_valve_signal)
-        }
-    }
-    impl ControllerSignal<TargetPressureTemperatureSignal> for TestApu {
-        fn signal(&self) -> Option<TargetPressureTemperatureSignal> {
-            Some(TargetPressureTemperatureSignal::new(
-                self.bleed_air_pressure,
-                self.bleed_air_temperature,
-            ))
-        }
-    }
-
-    struct TestEngineFirePushButtons {
-        is_released: [bool; 2],
-    }
-    impl TestEngineFirePushButtons {
-        fn new() -> Self {
-            Self {
-                is_released: [false, false],
-            }
-        }
-
-        fn release(&mut self, engine_number: usize) {
-            self.is_released[engine_number - 1] = true;
-        }
-    }
-    impl EngineFirePushButtons for TestEngineFirePushButtons {
-        fn is_released(&self, engine_number: usize) -> bool {
-            self.is_released[engine_number - 1]
-        }
-    }
-
-    struct A320TestElectrical {
-        airspeed: Velocity,
-        all_ac_lost: bool,
-    }
-    impl A320TestElectrical {
-        pub fn new() -> Self {
-            A320TestElectrical {
-                airspeed: Velocity::new::<knot>(100.),
-                all_ac_lost: false,
-            }
-        }
-
-        fn update(&mut self, context: &UpdateContext) {
-            self.airspeed = context.indicated_airspeed();
-        }
-    }
-    impl EmergencyElectricalState for A320TestElectrical {
-        fn is_in_emergency_elec(&self) -> bool {
-            self.all_ac_lost && self.airspeed >= Velocity::new::<knot>(100.)
-        }
-    }
-    impl SimulationElement for A320TestElectrical {
-        fn receive_power(&mut self, buses: &impl ElectricalBuses) {
-            self.all_ac_lost = !buses.is_powered(ElectricalBusType::AlternatingCurrent(1))
-                && !buses.is_powered(ElectricalBusType::AlternatingCurrent(2));
-        }
-    }
-
-    struct PneumaticTestAircraft {
-        pneumatic: A320Pneumatic,
-        air_conditioning: TestAirConditioning,
-        apu: TestApu,
-        engine_1: LeapEngine,
-        engine_2: LeapEngine,
-        pneumatic_overhead_panel: A320PneumaticOverheadPanel,
-        fire_pushbuttons: TestEngineFirePushButtons,
-        electrical: A320TestElectrical,
-        powered_source: TestElectricitySource,
-        dc_1_bus: ElectricalBus,
-        dc_2_bus: ElectricalBus,
-        dc_ess_bus: ElectricalBus,
-        dc_ess_shed_bus: ElectricalBus,
-        // Electric buses states to be able to kill them dynamically
-        is_dc_1_powered: bool,
-        is_dc_2_powered: bool,
-        is_dc_ess_powered: bool,
-        is_dc_ess_shed_powered: bool,
-    }
-    impl PneumaticTestAircraft {
-        fn new(context: &mut InitContext) -> Self {
-            Self {
-                pneumatic: A320Pneumatic::new(context),
-                air_conditioning: TestAirConditioning::new(context),
-                apu: TestApu::new(),
-                engine_1: LeapEngine::new(context, 1),
-                engine_2: LeapEngine::new(context, 2),
-                pneumatic_overhead_panel: A320PneumaticOverheadPanel::new(context),
-                fire_pushbuttons: TestEngineFirePushButtons::new(),
-                electrical: A320TestElectrical::new(),
-                powered_source: TestElectricitySource::powered(
-                    context,
-                    PotentialOrigin::EngineGenerator(1),
-                ),
-                dc_1_bus: ElectricalBus::new(context, ElectricalBusType::DirectCurrent(1)),
-                dc_2_bus: ElectricalBus::new(context, ElectricalBusType::DirectCurrent(2)),
-                dc_ess_bus: ElectricalBus::new(context, ElectricalBusType::DirectCurrentEssential),
-                dc_ess_shed_bus: ElectricalBus::new(
-                    context,
-                    ElectricalBusType::DirectCurrentEssentialShed,
-                ),
-                is_dc_1_powered: true,
-                is_dc_2_powered: true,
-                is_dc_ess_powered: true,
-                is_dc_ess_shed_powered: true,
-            }
-        }
-
-        fn set_dc_2_bus_power(&mut self, is_powered: bool) {
-            self.is_dc_2_powered = is_powered;
-        }
-
-        fn set_dc_ess_shed_bus_power(&mut self, is_powered: bool) {
-            self.is_dc_ess_shed_powered = is_powered;
-        }
-    }
-    impl Aircraft for PneumaticTestAircraft {
-        fn update_before_power_distribution(
-            &mut self,
-            _context: &UpdateContext,
-            electricity: &mut Electricity,
-        ) {
-            electricity.supplied_by(&self.powered_source);
-
-            if self.is_dc_1_powered {
-                electricity.flow(&self.powered_source, &self.dc_1_bus);
-            }
-
-            if self.is_dc_2_powered {
-                electricity.flow(&self.powered_source, &self.dc_2_bus);
-            }
-
-            if self.is_dc_ess_powered {
-                electricity.flow(&self.powered_source, &self.dc_ess_bus);
-            }
-
-            if self.is_dc_ess_shed_powered {
-                electricity.flow(&self.powered_source, &self.dc_ess_shed_bus);
-            }
-        }
-
-        fn update_after_power_distribution(&mut self, context: &UpdateContext) {
-            self.electrical.update(context);
-
-            self.apu.update(self.pneumatic.apu_bleed_air_valve());
-            self.pneumatic.update(
-                context,
-                [&self.engine_1, &self.engine_2],
-                &self.pneumatic_overhead_panel,
-                &self.fire_pushbuttons,
-                &self.apu,
-                &self.air_conditioning,
-            );
-            self.air_conditioning.update(
-                context,
-                [&self.engine_1, &self.engine_2],
-                &self.fire_pushbuttons,
-                &self.pneumatic,
-                &self.pneumatic_overhead_panel,
-            )
-        }
-    }
-    impl SimulationElement for PneumaticTestAircraft {
-        fn accept<T: SimulationElementVisitor>(&mut self, visitor: &mut T) {
-            self.electrical.accept(visitor);
-            self.pneumatic.accept(visitor);
-            self.engine_1.accept(visitor);
-            self.engine_2.accept(visitor);
-            self.pneumatic_overhead_panel.accept(visitor);
-            self.air_conditioning.accept(visitor);
-
-            visitor.visit(self);
-        }
-    }
-    struct PneumaticTestBed {
-        test_bed: SimulationTestBed<PneumaticTestAircraft>,
-    }
-    impl TestBed for PneumaticTestBed {
-        type Aircraft = PneumaticTestAircraft;
-
-        fn test_bed(&self) -> &SimulationTestBed<PneumaticTestAircraft> {
-            &self.test_bed
-        }
-
-        fn test_bed_mut(&mut self) -> &mut SimulationTestBed<PneumaticTestAircraft> {
-            &mut self.test_bed
-        }
-    }
-
-    impl PneumaticTestBed {
-        fn new() -> Self {
-            let mut test_bed = Self {
-                test_bed: SimulationTestBed::<PneumaticTestAircraft>::new(|context| {
-                    PneumaticTestAircraft::new(context)
-                }),
-            };
-            test_bed.command_pack_flow_selector_position(1);
-
-            //test_bed.command_pack_1_pb_position(true);
-            //test_bed.command_pack_2_pb_position(true);
-
-            test_bed
-        }
-
-        fn and_run(mut self) -> Self {
-            self.run();
-
-            self
-        }
-
-        fn and_stabilize(mut self) -> Self {
-            for _ in 1..1000 {
-                self.run_with_delta(Duration::from_millis(16));
-            }
-
-            self
-        }
-
-        fn and_stabilize_steps(mut self, n: usize) -> Self {
-            for _ in 1..n {
-                for _ in 1..1000 {
-                    self.run_with_delta(Duration::from_millis(16));
-                }
-            }
-            self
-        }
-
-        fn in_isa_atmosphere(mut self, altitude: Length) -> Self {
-            self.set_ambient_pressure(InternationalStandardAtmosphere::pressure_at_altitude(
-                altitude,
-            ));
-            self.set_ambient_temperature(InternationalStandardAtmosphere::temperature_at_altitude(
-                altitude,
-            ));
-
-            self
-        }
-
-        fn stop_eng1(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:1", false);
-            self.write_by_name("TURB ENG CORRECTED N1:1", Ratio::new::<ratio>(0.));
-            self.write_by_name("TURB ENG CORRECTED N2:1", Ratio::new::<ratio>(0.));
-            self.write_by_name("ENGINE_STATE:1", EngineState::Off);
-
-            self
-        }
-        fn stop_eng2(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:2", false);
-            self.write_by_name("TURB ENG CORRECTED N2:2", Ratio::new::<ratio>(0.));
-            self.write_by_name("TURB ENG CORRECTED N1:2", Ratio::new::<ratio>(0.));
-            self.write_by_name("ENGINE_STATE:2", EngineState::Off);
-
-            self
-        }
-
-        fn idle_eng1(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:1", true);
-            self.write_by_name("TURB ENG CORRECTED N2:1", Ratio::new::<ratio>(0.55));
-            self.write_by_name("TURB ENG CORRECTED N1:1", Ratio::new::<ratio>(0.2));
-            self.write_by_name("ENGINE_STATE:1", EngineState::On);
-
-            self
-        }
-
-        fn idle_eng2(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:2", true);
-            self.write_by_name("TURB ENG CORRECTED N2:2", Ratio::new::<ratio>(0.55));
-            self.write_by_name("TURB ENG CORRECTED N1:2", Ratio::new::<ratio>(0.2));
-            self.write_by_name("ENGINE_STATE:2", EngineState::On);
-
-            self
-        }
-
-        fn holding_eng1(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:1", true);
-            self.write_by_name("TURB ENG CORRECTED N2:1", Ratio::new::<ratio>(0.92));
-            self.write_by_name("TURB ENG CORRECTED N1:1", Ratio::new::<ratio>(0.5));
-            self.write_by_name("ENGINE_STATE:1", EngineState::On);
-
-            self
-        }
-
-        fn holding_eng2(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:2", true);
-            self.write_by_name("TURB ENG CORRECTED N2:2", Ratio::new::<ratio>(0.92));
-            self.write_by_name("TURB ENG CORRECTED N1:2", Ratio::new::<ratio>(0.5));
-            self.write_by_name("ENGINE_STATE:2", EngineState::On);
-
-            self
-        }
-
-        fn power_eng1(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:1", true);
-            self.write_by_name("TURB ENG CORRECTED N2:1", Ratio::new::<ratio>(0.65));
-            self.write_by_name("TURB ENG CORRECTED N1:1", Ratio::new::<ratio>(0.99));
-            self.write_by_name("ENGINE_STATE:1", EngineState::On);
-
-            self
-        }
-
-        fn power_eng2(mut self) -> Self {
-            self.write_by_name("GENERAL ENG STARTER ACTIVE:2", true);
-            self.write_by_name("TURB ENG CORRECTED N2:2", Ratio::new::<ratio>(0.65));
-            self.write_by_name("TURB ENG CORRECTED N1:2", Ratio::new::<ratio>(0.99));
-            self.write_by_name("ENGINE_STATE:2", EngineState::On);
-
-            self
-        }
-
-        fn wing_anti_ice_push_button(mut self, mode: WingAntiIcePushButtonMode) -> Self {
-            match mode {
-                WingAntiIcePushButtonMode::On => {
-                    self.write_by_name("BUTTON_OVHD_ANTI_ICE_WING_Position", true)
-                }
-                _ => self.write_by_name("BUTTON_OVHD_ANTI_ICE_WING_Position", false),
-            };
-
-            self
-        }
-
-        fn command_pack_flow_selector_position(&mut self, value: u8) {
-            self.write_by_name("KNOB_OVHD_AIRCOND_PACKFLOW_Position", value);
-        }
-
-        fn command_pack_1_pb_position(&mut self, value: bool) {
-            self.write_by_name("OVHD_COND_PACK_1_PB_IS_ON", value);
-        }
-
-        fn command_pack_2_pb_position(&mut self, value: bool) {
-            self.write_by_name("OVHD_COND_PACK_2_PB_IS_ON", value);
-        }
-
-        fn wing_anti_ice_system_on(&mut self) -> bool {
-            self.read_by_name("PNEU_WING_ANTI_ICE_SYSTEM_ON")
-        }
-
-        fn wing_anti_ice_system_selected(&mut self) -> bool {
-            self.read_by_name("PNEU_WING_ANTI_ICE_SYSTEM_SELECTED")
-        }
-
-        fn is_sim_on_ground(&mut self) -> bool {
-            self.read_by_name("SIM ON GROUND")
-        }
-
-        fn wing_anti_ice_has_fault(&mut self) -> bool {
-            self.read_by_name("PNEU_WING_ANTI_ICE_HAS_FAULT")
-        }
-
-        fn pack_1_has_fault(&mut self) -> bool {
-            self.read_by_name("OVHD_COND_PACK_1_PB_HAS_FAULT")
-        }
-
-        fn pack_1_on(&mut self) -> bool {
-            self.read_by_name("OVHD_COND_PACK_2_PB_IS_ON")
-        }
-
-        //Utility functions to get info from the test bed
-        fn left_precooler_pressurised(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.is_precoooler_pressurised(0))
-        }
-
-        fn right_precooler_pressurised(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.is_precoooler_pressurised(1))
-        }
-
-        fn left_wai_pressure(&self) -> Pressure {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_consumer_pressure(0))
-        }
-        fn right_wai_pressure(&self) -> Pressure {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_consumer_pressure(1))
-        }
-
-        fn left_wai_temperature(&self) -> ThermodynamicTemperature {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_consumer_temperature(0))
-        }
-
-        fn right_wai_temperature(&self) -> ThermodynamicTemperature {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_consumer_temperature(1))
-        }
-
-        fn left_precooler_pressure(&self) -> Pressure {
-            self.query(|a| a.pneumatic.engine_systems[0].precooler_outlet_pressure())
-        }
-
-        fn right_precooler_pressure(&self) -> Pressure {
-            self.query(|a| a.pneumatic.engine_systems[1].precooler_outlet_pressure())
-        }
-
-        fn left_precooler_temperature(&self) -> ThermodynamicTemperature {
-            self.query(|a| a.pneumatic.engine_systems[0].precooler_outlet_temperature())
-        }
-
-        fn right_precooler_temperature(&self) -> ThermodynamicTemperature {
-            self.query(|a| a.pneumatic.engine_systems[1].precooler_outlet_temperature())
-        }
-
-        fn left_valve_open_amount(&self) -> f64 {
-            self.query(|a| {
-                a.pneumatic.wing_anti_ice.wai_valve[0]
-                    .open_amount()
-                    .get::<ratio>()
-            })
-        }
-
-        fn right_valve_open_amount(&self) -> f64 {
-            self.query(|a| {
-                a.pneumatic.wing_anti_ice.wai_valve[1]
-                    .open_amount()
-                    .get::<ratio>()
-            })
-        }
-
-        fn left_valve_controller_timer(&self) -> Duration {
-            self.query(|a| a.pneumatic.wing_anti_ice.valve_controller[0].get_timer())
-        }
-
-        fn right_valve_controller_timer(&self) -> Duration {
-            self.query(|a| a.pneumatic.wing_anti_ice.valve_controller[1].get_timer())
-        }
-
-        fn left_valve_pid_output(&self) -> f64 {
-            self.query(|a| {
-                a.pneumatic.wing_anti_ice.valve_controller[0].controller_valve_pid_output()
-            })
-        }
-
-        fn right_valve_pid_output(&self) -> f64 {
-            self.query(|a| {
-                a.pneumatic.wing_anti_ice.valve_controller[1].controller_valve_pid_output()
-            })
-        }
-
-        fn left_valve_controller_on(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.valve_controller[0].controller_signals_on())
-        }
-
-        fn right_valve_controller_on(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.valve_controller[1].controller_signals_on())
-        }
-
-        fn left_valve_closed(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.is_wai_valve_closed(0))
-        }
-
-        fn right_valve_closed(&self) -> bool {
-            self.query(|a| a.pneumatic.wing_anti_ice.is_wai_valve_closed(1))
-        }
-
-        fn left_valve_open(&self) -> bool {
-            !self.left_valve_closed()
-        }
-
-        fn right_valve_open(&self) -> bool {
-            !self.right_valve_closed()
-        }
-
-        fn left_exhaust_flow(&self) -> MassRate {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_exhaust[0].fluid_flow())
-        }
-
-        fn right_exhaust_flow(&self) -> MassRate {
-            self.query(|a| a.pneumatic.wing_anti_ice.wai_exhaust[1].fluid_flow())
-        }
-    }
-
-    fn test_bed() -> PneumaticTestBed {
-        PneumaticTestBed::new()
-    }
-
-    #[test]
-    fn wing_anti_ice_simvars() {
-        let test_bed = test_bed();
-
-        assert!(test_bed.contains_variable_with_name("PNEU_WING_ANTI_ICE_SYSTEM_ON"));
-        assert!(test_bed.contains_variable_with_name("PNEU_WING_ANTI_ICE_SYSTEM_SELECTED"));
-        assert!(test_bed.contains_variable_with_name("PNEU_WING_ANTI_ICE_HAS_FAULT"));
-        assert!(test_bed.contains_variable_with_name("PNEU_1_WING_ANTI_ICE_CONSUMER_PRESSURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_2_WING_ANTI_ICE_CONSUMER_PRESSURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_1_WING_ANTI_ICE_CONSUMER_TEMPERATURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_2_WING_ANTI_ICE_CONSUMER_TEMPERATURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_1_WING_ANTI_ICE_VALVE_CLOSED"));
-        assert!(test_bed.contains_variable_with_name("PNEU_2_WING_ANTI_ICE_VALVE_CLOSED"));
-        assert!(test_bed.contains_variable_with_name("BUTTON_OVHD_ANTI_ICE_WING_Position"));
-
-        assert!(test_bed.contains_variable_with_name("PNEU_1_WING_ANTI_ICE_HIGH_PRESSURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_2_WING_ANTI_ICE_HIGH_PRESSURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_1_WING_ANTI_ICE_LOW_PRESSURE"));
-        assert!(test_bed.contains_variable_with_name("PNEU_2_WING_ANTI_ICE_LOW_PRESSURE"));
-    }
-
-    #[test]
-    fn wing_anti_ice_cold_and_dark() {
-        let altitude = Length::new::<foot>(500.);
-        let ambient_pressure = InternationalStandardAtmosphere::pressure_at_altitude(altitude);
-        let ambient_temperature =
-            InternationalStandardAtmosphere::temperature_at_altitude(altitude);
-
-        let temperature_epsilon = ThermodynamicTemperature::new::<degree_celsius>(0.05);
-        let pressure_epsilon = Pressure::new::<psi>(0.01);
-
-        let mut test_bed = test_bed()
-            .stop_eng1()
-            .stop_eng2()
-            .in_isa_atmosphere(altitude);
-        test_bed.set_on_ground(true);
-        test_bed = test_bed.and_stabilize_steps(7);
-
-        println!("left press = {}", test_bed.left_wai_pressure().get::<psi>());
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-        println!("ambient press = {}", ambient_pressure.get::<psi>());
-        println!(
-            "left temp = {}",
-            test_bed.left_wai_temperature().get::<degree_celsius>()
-        );
-        println!(
-            "right temp = {}",
-            test_bed.right_wai_temperature().get::<degree_celsius>()
-        );
-        println!(
-            "ambient temp = {}",
-            ambient_temperature.get::<degree_celsius>()
-        );
-
-        assert!((test_bed.left_wai_pressure() - ambient_pressure).abs() < pressure_epsilon);
-        assert!((test_bed.right_wai_pressure() - ambient_pressure).abs() < pressure_epsilon);
-        assert!(
-            (test_bed.left_wai_temperature().get::<degree_celsius>()
-                - ambient_temperature.get::<degree_celsius>())
-            .abs()
-                < temperature_epsilon.get::<degree_celsius>()
-        );
-        assert!(
-            (test_bed.right_wai_temperature().get::<degree_celsius>()
-                - ambient_temperature.get::<degree_celsius>())
-            .abs()
-                < temperature_epsilon.get::<degree_celsius>()
-        );
-        assert!(!test_bed.left_valve_open());
-        assert!(!test_bed.right_valve_open());
-        assert!(!test_bed.wing_anti_ice_system_on());
-        assert!(!test_bed.wing_anti_ice_has_fault());
-    }
-
-    #[test]
-    fn wing_anti_ice_on_and_off() {
-        let altitude = Length::new::<foot>(20000.);
-        let ambient_pressure = InternationalStandardAtmosphere::pressure_at_altitude(altitude);
-        let ambient_temperature =
-            InternationalStandardAtmosphere::temperature_at_altitude(altitude);
-
-        let temperature_epsilon = ThermodynamicTemperature::new::<degree_celsius>(0.05);
-        let pressure_epsilon = Pressure::new::<psi>(0.01);
-
-        let mut test_bed = test_bed()
-            .power_eng1()
-            .power_eng2()
-            .in_isa_atmosphere(altitude);
-        test_bed.set_on_ground(false);
-        test_bed = test_bed.and_stabilize_steps(7);
-
-        test_bed = test_bed
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::Off)
-            .and_stabilize();
-
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-        println!("ambient press = {}", ambient_pressure.get::<psi>());
-
-        assert!((test_bed.left_wai_pressure() - ambient_pressure).abs() < pressure_epsilon);
-        assert!((test_bed.right_wai_pressure() - ambient_pressure).abs() < pressure_epsilon);
-        assert!(
-            (test_bed.left_wai_temperature().get::<degree_celsius>()
-                - ambient_temperature.get::<degree_celsius>())
-            .abs()
-                < temperature_epsilon.get::<degree_celsius>()
-        );
-        assert!(
-            (test_bed.right_wai_temperature().get::<degree_celsius>()
-                - ambient_temperature.get::<degree_celsius>())
-            .abs()
-                < temperature_epsilon.get::<degree_celsius>()
-        );
-        assert!(!test_bed.left_valve_open());
-        assert!(!test_bed.right_valve_open());
-        assert!(!test_bed.wing_anti_ice_system_on());
-        assert!(!test_bed.wing_anti_ice_has_fault());
-
-        test_bed = test_bed
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize();
-
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-
-        println!(
-            "right precooler press = {}",
-            test_bed.right_precooler_pressure().get::<psi>()
-        );
-
-        // assert!(
-        //     (test_bed.left_wai_pressure() - test_bed.precooler_pressure(0)).abs()
-        //         < pressure_epsilon
-        // );
-        // assert!(
-        //     (test_bed.right_wai_pressure() - test_bed.precooler_pressure(1)).abs()
-        //         < pressure_epsilon
-        // );
-
-        assert!(test_bed.left_valve_open());
-        assert!(test_bed.right_valve_open());
-        assert!(test_bed.wing_anti_ice_system_on());
-        assert!(!test_bed.wing_anti_ice_has_fault());
-    }
-
-    #[test]
-    fn wing_anti_ice_two_engine_idle_full_state() {
-        let altitude = Length::new::<foot>(0.);
-        let test_bed = test_bed()
-            .idle_eng1()
-            .idle_eng2()
-            .in_isa_atmosphere(altitude)
-            //.mach_number(MachNumber(0.))
-            .and_stabilize();
-
-        println!(
-            "left precooler press = {}",
-            test_bed.left_precooler_pressure().get::<psi>()
-        );
-
-        assert!(false);
-    }
-
-    #[test]
-    fn wing_anti_ice_has_fault_when_precooler_not_pressurized() {
-        let altitude = Length::new::<foot>(500.);
-
-        let mut test_bed = test_bed()
-            .stop_eng1()
-            .stop_eng2()
-            .in_isa_atmosphere(altitude);
-        test_bed.set_on_ground(true);
-        test_bed = test_bed.and_stabilize();
-
-        test_bed = test_bed
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize();
-
-        println!("left open amount = {}", test_bed.left_valve_open_amount());
-        println!("right open amount = {}", test_bed.right_valve_open_amount());
-
-        println!(
-            "left precooler press = {}",
-            test_bed.left_precooler_pressure().get::<psi>()
-        );
-        println!(
-            "right precooler press = {}",
-            test_bed.right_precooler_pressure().get::<psi>()
-        );
-        assert!(test_bed.wing_anti_ice_has_fault());
-    }
-
-    #[test]
-    fn wing_anti_ice_no_fault_after_starting_engine() {
-        let altitude = Length::new::<foot>(500.);
-        let _wai_pressure: Pressure = Pressure::new::<psi>(22.5);
-        let _pressure_epsilon: Pressure = Pressure::new::<psi>(0.1);
-
-        let mut test_bed = test_bed()
-            .stop_eng1()
-            .stop_eng2()
-            .in_isa_atmosphere(altitude);
-        test_bed.set_on_ground(true);
-        test_bed = test_bed.and_stabilize();
-
-        test_bed = test_bed
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize();
-        assert!(test_bed.wing_anti_ice_has_fault());
-
-        test_bed = test_bed.idle_eng1().idle_eng2().and_stabilize();
-        println!("left press = {}", test_bed.left_wai_pressure().get::<psi>());
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-        println!(
-            "left precooler pressurised = {}",
-            test_bed.left_precooler_pressurised()
-        );
-        println!(
-            "right precooler pressurised = {}",
-            test_bed.right_precooler_pressurised()
-        );
-        println!(
-            "left controller_signals_on = {}",
-            test_bed.left_valve_controller_on()
-        );
-        println!(
-            "right controller_signals_on = {}",
-            test_bed.right_valve_controller_on()
-        );
-        println!("left open amount = {}", test_bed.left_valve_open_amount());
-        println!("right open amount = {}", test_bed.right_valve_open_amount());
-        println!(
-            "left valve_pid_output = {}",
-            test_bed.left_valve_pid_output()
-        );
-        println!(
-            "right valve_pid_output = {}",
-            test_bed.right_valve_pid_output()
-        );
-        assert!(!test_bed.wing_anti_ice_has_fault());
-        assert!(test_bed.wing_anti_ice_system_on());
-    }
-
-    #[test]
-    fn wing_anti_ice_timer_doesnt_start_if_precooler_not_pressurized() {
-        let altitude = Length::new::<foot>(500.);
-
-        let mut test_bed = test_bed()
-            .stop_eng1()
-            .stop_eng2()
-            .in_isa_atmosphere(altitude);
-        test_bed.set_on_ground(true);
-        test_bed = test_bed.and_stabilize();
-
-        test_bed = test_bed
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize();
-        assert!(test_bed.wing_anti_ice_has_fault());
-        assert!(test_bed.left_valve_controller_timer() == Duration::from_secs(0));
-        assert!(test_bed.right_valve_controller_timer() == Duration::from_secs(0));
-    }
-
-    #[test]
-    fn wing_anti_ice_tweak_pid() {
-        let altitude = Length::new::<foot>(10.);
-        let _wai_pressure: Pressure = Pressure::new::<psi>(22.5);
-        let _pressure_epsilon: Pressure = Pressure::new::<psi>(0.1);
-        let mut test_bed = test_bed()
-            .in_isa_atmosphere(altitude)
-            .idle_eng1()
-            .idle_eng2()
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On);
-        test_bed.set_on_ground(false);
-        test_bed = test_bed.and_stabilize();
-
-        assert!(test_bed.left_valve_controller_on());
-
-        for _ in 0..50 {
-            println!(
-                "left pressure = {}",
-                test_bed.left_wai_pressure().get::<psi>()
-            );
-            println!("left open amount = {}", test_bed.left_valve_open_amount());
-            assert!(test_bed.left_valve_open_amount() != 0.);
-            test_bed.run_with_delta(Duration::from_millis(16));
-        }
-
-        assert!(test_bed.left_valve_open());
-    }
-
-    #[test]
-    fn wing_anti_ice_time_to_open() {
-        let altitude = Length::new::<foot>(2000.);
-        let wai_pressure: Pressure = Pressure::new::<psi>(22.5);
-        let pressure_epsilon: Pressure = Pressure::new::<psi>(0.1);
-        let mut test_bed = test_bed()
-            .in_isa_atmosphere(altitude)
-            .holding_eng1()
-            .holding_eng2();
-        test_bed.set_on_ground(false);
-        test_bed = test_bed.and_stabilize_steps(10);
-        test_bed = test_bed.wing_anti_ice_push_button(WingAntiIcePushButtonMode::On);
-        test_bed.run_with_delta(Duration::from_millis(16));
-
-        let mut time_to_open_valve = Duration::from_millis(0);
-        for _ in 0..10000 {
-            // println!(
-            //     "left pressure = {}",
-            //     test_bed.left_wai_pressure().get::<psi>()
-            // );
-            // println!(
-            //     "left precooler pressurised = {}",
-            //     test_bed.left_precooler_pressurised()
-            // );
-            // println!(
-            //     "left precooler press = {}",
-            //     test_bed.left_precooler_pressure().get::<psi>()
-            // );
-            // println!(
-            //     "left exhaust flow = {}",
-            //     test_bed.left_exhaust_flow().get::<kilogram_per_second>()
-            // );
-            //println!("-----------------------------------------",);
-            // test_bed.left_wai_pressure().get::<psi>() < 22.;
-            let wai_has_fault = test_bed.wing_anti_ice_has_fault();
-            if wai_has_fault {
-                time_to_open_valve += Duration::from_millis(16);
-            } else {
-                //break;
-            }
-            test_bed.run_with_delta(Duration::from_millis(16));
-        }
-        println!(
-            "left pressure = {}",
-            test_bed.left_wai_pressure().get::<psi>()
-        );
-        println!(
-            "left precooler pressurised = {}",
-            test_bed.left_precooler_pressurised()
-        );
-        println!(
-            "left precooler press = {}",
-            test_bed.left_precooler_pressure().get::<psi>()
-        );
-        println!(
-            "left exhaust flow = {}",
-            test_bed.left_exhaust_flow().get::<kilogram_per_second>()
-        );
-        println!("left pack fault = {}", test_bed.pack_1_has_fault());
-        println!("left pack on = {}", test_bed.pack_1_on());
-        println!(
-            "Time to open left valve = {}",
-            time_to_open_valve.as_secs_f32()
-        );
-        //assert!(test_bed.left_wai_pressure().get::<psi>() >= 22.);
-        assert!(time_to_open_valve.as_secs_f32() >= 2.);
-        assert!(time_to_open_valve.as_secs_f32() <= 3.);
-        assert!(!test_bed.left_valve_open());
-    }
-
-    #[test]
-    fn wing_anti_ice_tweak_exhaust() {
-        let altitude = Length::new::<foot>(22000.);
-        let ambient_pressure = InternationalStandardAtmosphere::pressure_at_altitude(altitude);
-        let ambient_temperature =
-            InternationalStandardAtmosphere::temperature_at_altitude(altitude);
-
-        let _wai_pressure: Pressure = Pressure::new::<psi>(22.5);
-        let _pressure_epsilon: Pressure = Pressure::new::<psi>(0.1);
-        let test_bed = test_bed()
-            .in_isa_atmosphere(altitude)
-            .power_eng1()
-            .power_eng2()
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize_steps(10);
-
-        assert!(test_bed.left_valve_controller_on());
-        assert!(test_bed.left_valve_open());
-
-        let left_flow_rate = test_bed.left_exhaust_flow().get::<kilogram_per_second>();
-        let right_flow_rate = test_bed.right_exhaust_flow().get::<kilogram_per_second>();
-
-        println!("left press = {}", test_bed.left_wai_pressure().get::<psi>());
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-        println!("ambient press = {}", ambient_pressure.get::<psi>());
-        println!(
-            "left temp = {}",
-            test_bed.left_wai_temperature().get::<degree_celsius>()
-        );
-        println!(
-            "right temp = {}",
-            test_bed.right_wai_temperature().get::<degree_celsius>()
-        );
-        println!(
-            "ambient temp = {}",
-            ambient_temperature.get::<degree_celsius>()
-        );
-        println!("left exhaust flow = {}", left_flow_rate);
-        println!("right exhaust flow = {}", right_flow_rate);
-
-        // At 22000ft, flow rate is given at 0.327kg/s
-        assert!(((left_flow_rate * 1000.0).round() - 327.0).abs() < 1.);
-        assert!(((right_flow_rate * 1000.0).round() - 327.0).abs() < 1.);
-    }
-
-    #[test]
-    fn wing_anti_ice_pressure_regulated() {
-        let altitude = Length::new::<foot>(6000.);
-        let wai_pressure: Pressure = Pressure::new::<psi>(22.5);
-        let pressure_epsilon: Pressure = Pressure::new::<psi>(0.1);
-
-        let test_bed = test_bed()
-            .in_isa_atmosphere(altitude)
-            .power_eng1()
-            .power_eng2()
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize_steps(10);
-
-        println!(
-            "left pressure = {}",
-            test_bed.left_wai_pressure().get::<psi>()
-        );
-        println!(
-            "right pressure = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-
-        assert!((test_bed.left_wai_pressure() - wai_pressure).abs() < pressure_epsilon);
-        assert!((test_bed.right_wai_pressure() - wai_pressure).abs() < pressure_epsilon);
-    }
-
-    #[test]
-    fn wing_anti_ice_valve_close_after_30_seconds_on_ground() {
-        let mut test_bed = test_bed()
-            .idle_eng1()
-            .idle_eng2()
-            .in_isa_atmosphere(Length::new::<foot>(0.))
-            .and_stabilize();
-        test_bed.set_on_ground(true);
-
-        assert!(test_bed.left_valve_controller_timer() == Duration::from_secs(0));
-        assert!(test_bed.right_valve_controller_timer() == Duration::from_secs(0));
-
-        test_bed = test_bed.wing_anti_ice_push_button(WingAntiIcePushButtonMode::On);
-        test_bed.run_with_delta(Duration::from_millis(16));
-        assert!(test_bed.wing_anti_ice_system_on());
-        test_bed.run_with_delta(Duration::from_secs(1));
-
-        assert!(test_bed.left_valve_open_amount() > 0.);
-        assert!(test_bed.left_valve_controller_timer() < Duration::from_secs(2));
-        assert!(test_bed.right_valve_open_amount() > 0.);
-        assert!(test_bed.right_valve_controller_timer() < Duration::from_secs(2));
-
-        test_bed.run_with_delta(Duration::from_secs(30));
-        assert!(test_bed.left_valve_controller_timer() == Duration::from_secs(30));
-        assert!(test_bed.left_valve_open_amount() == 0.);
-        assert!(test_bed.right_valve_controller_timer() == Duration::from_secs(30));
-        assert!(test_bed.right_valve_open_amount() == 0.);
-        assert!(!test_bed.wing_anti_ice_system_on());
-        assert!(test_bed.wing_anti_ice_system_selected());
-    }
-
-    #[test]
-    fn wing_anti_ice_valve_open_after_leaving_ground_after_test() {
-        let _altitude = Length::new::<foot>(500.);
-        let mut test_bed = test_bed()
-            .idle_eng1()
-            .idle_eng2()
-            .in_isa_atmosphere(Length::new::<foot>(0.))
-            .and_stabilize();
-        test_bed.set_on_ground(true);
-
-        test_bed = test_bed.wing_anti_ice_push_button(WingAntiIcePushButtonMode::On);
-        test_bed.run_with_delta(Duration::from_secs(31));
-
-        assert!(!test_bed.wing_anti_ice_system_on());
-        assert!(test_bed.wing_anti_ice_system_selected());
-        assert!(test_bed.left_valve_open_amount() == 0.);
-        assert!(test_bed.right_valve_open_amount() == 0.);
-
-        test_bed.set_on_ground(false);
-        test_bed.run_with_delta(Duration::from_secs(1));
-
-        println!("left press = {}", test_bed.left_wai_pressure().get::<psi>());
-        println!(
-            "right press = {}",
-            test_bed.right_wai_pressure().get::<psi>()
-        );
-        println!(
-            "left precooler pressurised = {}",
-            test_bed.left_precooler_pressurised()
-        );
-        println!(
-            "right precooler pressurised = {}",
-            test_bed.right_precooler_pressurised()
-        );
-        println!(
-            "left controller_signals_on = {}",
-            test_bed.left_valve_controller_on()
-        );
-        println!(
-            "right controller_signals_on = {}",
-            test_bed.right_valve_controller_on()
-        );
-        println!("left open amount = {}", test_bed.left_valve_open_amount());
-        println!("right open amount = {}", test_bed.right_valve_open_amount());
-        println!(
-            "left valve_pid_output = {}",
-            test_bed.left_valve_pid_output()
-        );
-        println!(
-            "right valve_pid_output = {}",
-            test_bed.right_valve_pid_output()
-        );
-
-        assert!(test_bed.wing_anti_ice_system_on());
-        assert!(test_bed.left_valve_open_amount() > 0.);
-        assert!(test_bed.right_valve_open_amount() > 0.);
-        assert!(test_bed.left_valve_controller_timer() == Duration::from_secs(0));
-        assert!(test_bed.right_valve_controller_timer() == Duration::from_secs(0));
-    }
-
-    #[test]
-    fn wing_anti_ice_valve_controller_timer_starts_after_landing() {
-        let altitude = Length::new::<foot>(500.);
-        let mut test_bed = test_bed()
-            .idle_eng1()
-            .idle_eng2()
-            .in_isa_atmosphere(altitude)
-            .wing_anti_ice_push_button(WingAntiIcePushButtonMode::On)
-            .and_stabilize();
-
-        test_bed.set_on_ground(true);
-        test_bed.run_with_delta(Duration::from_secs(30));
-        assert!(!test_bed.left_valve_open());
-        assert!(!test_bed.right_valve_open());
-
-        test_bed.set_on_ground(false);
-        test_bed.run_with_delta(Duration::from_millis(16));
-
-        assert!(test_bed.left_valve_open_amount() > 0.);
-        assert!(test_bed.right_valve_open_amount() > 0.);
-        assert!(test_bed.left_valve_controller_timer() == Duration::from_secs(0));
-        assert!(test_bed.right_valve_controller_timer() == Duration::from_secs(0));
-
-        test_bed.set_on_ground(true);
-        test_bed.run_with_delta(Duration::from_millis(500));
-
-        assert!(test_bed.left_valve_open_amount() > 0.);
-        assert!(test_bed.right_valve_open_amount() > 0.);
-        assert!(test_bed.left_valve_controller_timer() > Duration::from_secs(0));
-        assert!(test_bed.right_valve_controller_timer() > Duration::from_secs(0));
-
-        test_bed.run_with_delta(Duration::from_secs(30));
-        assert!(test_bed.left_valve_open_amount() == 0.);
-        assert!(test_bed.right_valve_open_amount() == 0.);
     }
 }
